feat(gabinet): extract shared SettlementForm + wire into appointment detail (closes #3516)
- New src/components/gabinet/appointment-shared/settlement-form.tsx:
  canonical payment method list (cash/card/transfer/package/gratis/barter/other),
  amount due from junction treatments (priceAtBooking ?? catalog, legacy fallback),
  package picker filtered to active packages covering any visit treatment
  (multi-treatment aware), discount/credit/overpayment logic, submit
- Route _layout.gabinet.appointments.$appointmentId.lazy.tsx: replace ~300-line
  inline payment dialog with <SettlementForm>, remove 10 payment state vars,
  remove openPaymentDialog/handleCreatePayment functions, remove
  createPayment/getPatientCreditAction hook calls
- i18n: gabinet.payments.methods.package → PL "Karnet/seria/pakiet",
  EN "Package/Series/Carnet"
